[NN] Refactor the Code Structure of GT (#5100)
