+++ python/dgl/nn/pytorch/__init__.py
@@ -8,12 +8,6 @@ from .softmax import *
 from .factory import *
 from .hetero import *
 from .sparse_emb import NodeEmbedding
-from .utils import (
-    JumpingKnowledge,
-    LabelPropagation,
-    LaplacianPosEnc,
-    Sequential,
-    WeightBasis,
-)
+from .utils import JumpingKnowledge, LabelPropagation, Sequential, WeightBasis
 from .network_emb import *
-from .graph_transformer import *
+from .gt import *
