+++ python/dgl/transforms/module.py
@@ -19,7 +19,7 @@
 from scipy.linalg import expm
 
 from .. import backend as F, convert, function as fn, utils
-from ..base import DGLError
+from ..base import dgl_warning, DGLError
 from . import functional
 
 try:
@@ -34,6 +34,7 @@ __all__ = [
     "FeatMask",
     "RandomWalkPE",
     "LaplacianPE",
+    "LapPE",
     "AddSelfLoop",
     "RemoveSelfLoop",
     "AddReverse",
@@ -419,7 +420,7 @@ class RandomWalkPE(BaseTransform):
         return g
 
 
-class LaplacianPE(BaseTransform):
+class LapPE(BaseTransform):
     r"""Laplacian Positional Encoding, as introduced in
     `Benchmarking Graph Neural Networks
     <https://arxiv.org/abs/2003.00982>`__
@@ -433,23 +434,21 @@ class LaplacianPE(BaseTransform):
     feat_name : str, optional
         Name to store the computed positional encodings in ndata.
     eigval_name : str, optional
-        If None, store laplacian eigenvectors only.
-        Otherwise, it's the name to store corresponding laplacian eigenvalues in ndata.
-        Default: None.
+        If None, store laplacian eigenvectors only. Otherwise, it's the name to
+        store corresponding laplacian eigenvalues in ndata. Default: None.
     padding : bool, optional
         If False, raise an exception when k>=n.
-        Otherwise, add zero paddings in the end of eigenvectors and 'nan' paddings
-        in the end of eigenvalues when k>=n.
-        Default: False.
+        Otherwise, add zero paddings in the end of eigenvectors and 'nan'
+        paddings in the end of eigenvalues when k>=n. Default: False.
         n is the number of nodes in the given graph.
 
     Example
     -------
     >>> import dgl
-    >>> from dgl import LaplacianPE
-    >>> transform1 = LaplacianPE(k=3)
-    >>> transform2 = LaplacianPE(k=5, padding=True)
-    >>> transform3 = LaplacianPE(k=5, feat_name='eigvec', eigval_name='eigval', padding=True)
+    >>> from dgl import LapPE
+    >>> transform1 = LapPE(k=3)
+    >>> transform2 = LapPE(k=5, padding=True)
+    >>> transform3 = LapPE(k=5, feat_name='eigvec', eigval_name='eigval', padding=True)
     >>> g = dgl.graph(([0,1,2,3,4,2,3,1,4,0], [2,3,1,4,0,0,1,2,3,4]))
     >>> g1 = transform1(g)
     >>> print(g1.ndata['PE'])
@@ -488,16 +487,24 @@ class LaplacianPE(BaseTransform):
 
     def __call__(self, g):
         if self.eigval_name:
-            PE, eigval = functional.laplacian_pe(
+            PE, eigval = functional.lap_pe(
                 g, k=self.k, padding=self.padding, return_eigval=True
             )
             eigval = F.repeat(F.reshape(eigval, [1, -1]), g.num_nodes(), dim=0)
             g.ndata[self.eigval_name] = F.copy_to(eigval, g.device)
         else:
-            PE = functional.laplacian_pe(g, k=self.k, padding=self.padding)
+            PE = functional.lap_pe(g, k=self.k, padding=self.padding)
         g.ndata[self.feat_name] = F.copy_to(PE, g.device)
 
         return g
+
+
+class LaplacianPE(LapPE):
+    r"""Alias of `LapPE`."""
+
+    def __init__(self, k, feat_name="PE", eigval_name=None, padding=False):
+        super().__init__(k, feat_name, eigval_name, padding)
+        dgl_warning("LaplacianPE will be deprecated. Use LapPE please.")
 
 
 class AddSelfLoop(BaseTransform):
